Update how-to-access-api.md
Line 16.  "To get Access Token, App Key and App Secret on the Developer Website " chg to  "Get Access Token, App Key and App Secret on the Developer Website"
Line 32 "Set the request parameter header information, X-Api-Key, Authorization, X-Timestamp will be used in the signature function." - "Set the request parameter header information, and X-Api-Key, Authorization, X-Timestamp will be used in the signature function."

line67 " Signing the request and sets the signature in the request header X-Api-Signature. - Sign the request and set the signature in the request header X-Api-Signature."